Support for checking/getting packages by name/arch from a list or set
This adds new has_package and find_package methods to the
LCFGPackageList and LCFGPackageSet classes. They both take name and
(optionally) architecture strings as parameters. It also adds the
__contains__ and __getitem__ methods for convenience which will take
either just the name string or a tuple of name and architecture
strings.

A couple of small tests are included but the testing of these features
still needs more fleshing out.